Skeleton for C-interface ready. Currently only provides xSWAP
The intention is to provide a libviennacl.so (and a DLL on Windows)
so that the host-interface, CUDA, and OpenCL kernels are made available
to languages other than C++. The interface is BLAS-like, giving access to
either of the three backends. In addition, an object-oriented
backend-agnostic version of the BLAS functions is also provided.

Currently requires CUDA and OpenCL to be enabled. This is a leftover
from some fight involving CUDA, CMake, and myself. Will be rectified later.