Vector: CTOR of vector_base now creates a buffer copy.
Either vector_base/matrix_base always create shallow copies in their copy-CTOR,
or they create deep copies in their copy-CTOR. The side-effects of shallow
copies can be corrupted data, leading to wrong results. On the contrary,
deep copies are likely to have only poor performance as a side-effect,
which is not as bad. The only use for shallow copies is within
proxy objects, which now have suitable overloads for their copy-CTOR.

Resolves #60