AMG: Fixed incorrect coarse hierarchy construction for MIS-2.
Propagation of coarse indices was handled incorrectly, non-direct
neighbors did not get a coarse index assigned. This is resolved with
an additional 'reset' of the node state to 'undecided' if not a MIS node.
However, this requires another loop over all nodes plus a synchronization.

Also, note that the sequential version is not exactly the same as the MIS-2
construction because of the way non-neighbor nodes are aggregated.
Expect slight deviations in iteration counts (usually a few percent)
in addition to the randomization.