prod(): Added overloads for matrix_expression and vector_expression.
Allows users to write e.g. A = prod(B+C, x + y);
Any expressions passed to prod() are converted to temporaries.
This is certainly desirable from a performance point of view for GEMM.
One could do better with expression templates for GEMV, but that would not work with OpenCL.
Fixes #126.