Block-ILU: Removed unnecessary temporary matrices in setup stage.
Improves performance by about a factor of three to four.
Most of the optimization potential is now leveraged:
Each of the block matrices could also be removed by extending the
ILU0 and ILUT preconditioners accordingly. This, however, requires
larger refactoring and in view of the parallel Chow-Patel ILU
may not be worth the effort.

Resolves #40.