sum: Added support for sum(), row_sum(), column_sum()
These routines compute the sum of a vector as well as the row- and
column-sums of a dense matrix, respectively. The implementation
reuses inner products for vectors and matrix-vector products for
matrices. Thus, there is some overhead when compared to super-optimized
routines involved, but this should be acceptable in almost all cases.

Replaces former attempts of a reduce<>() function.

Resolves #127.