cache size information on AMD systems are now read correctly
The inline assembler in get_cache_sizes() gets a pointer to an array which should be stored in %rdi.
This was the only way I could get it to work propperly, as specifying input/output operands would yield segfaults.
Therefore, the inline assembler is in a seperate function and relies on the standard register or first function argument (i.e. %rdi).
I do not know if this could cause problems on other systems => needs to be tested.