cpuid is read correctly in either one of two ways on intel cpus, not thouroughly tested
CPUID info can be obtained through cpuid-leaf2 or cpuid-leaf4 on intel CPUs.
It depends on the CPU, which leaf to use. Both have been implemented and
leaf2 works correctly on a core 2 quad q9400. Further,thorough testing and double checking of the huge switch-case for leaf2 has NOT been done.